Task 04
Hexagon is a geometrical shape having six sides while square is a shape which has four equal sides. You

ane required to write Python or C++ program for hexagon anti square having data members and member functions, and create a repeating menu keeping in mind the following requirements.

1. Consider last digit of your CNIC as the length of one side of hexagon. For example, if your CNIC is XY210351532, the last digit will be 2.

2. Calculate area and perimeter of hexagon. (Hint: Area of hexagon 1.5*1.732*s where 's' is the length of one side of hexagon. Perimeter of hexagon 6*s; where's' is the length of one side of hexagon)

3. Calculate sum of all the angles of hexagon: (Hint: Sum of all the angles of hexagon 6a where 'a' is

the measurement of one angle of hexagon which is equal to 120.) 4. When the input of user is "T', display area, perimeter and sum of all the angles of hexagon.

5. Calculate area and perimeter of square whereas

a.Length of one side of square-last digit of CNIC + 1. (Hint: Area of square (length)2; where length is the length of one side of square. Perimeter of square= 4 Mength; where 'length' is the length of one side of the square)

6. When the input of user is 2, display the area and perimeter of the square.

7. On any other input, the program should exit

Instructions to write the program:

Use the last digit of your own CNIC If you do not use it, your marks will be deducted.

Write member functions to calculate and display area, perameter, and sum of angles of the hexagon. The following function names should be used for consistency in the hexagon class.